Revise copy/migration from sqlite to postgres
Change to ask for explicit inputs for the db dir and uri rather than
deriving the the uri from the dir.
